CHHYBRID-2711: Added the shutdown feature to the `MuleAgentAdministrationService`. Moved the exceptions to a common package. Changed the UnifiedRuntimeUtils from a helper to a Singleton in order to make it testable.
