+++ mule-agent-eventtracking-internalhandlers/mule-agent-db-event-tracker/src/test/java/com/mulesoft/agent/eventtracking/db/EventTrackingDBInternalHandlerTest.java
@@ -11,7 +11,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import com.mulesoft.agent.AgentEnableOperationException;
+import com.mulesoft.agent.exception.AgentEnableOperationException;
 import com.mulesoft.agent.domain.tracking.AgentTrackingNotification;
 
 import org.junit.Assert;
