+++ mule-agent-eventtracking-internalhandlers/mule-agent-splunk-event-tracker/src/test/java/com/mulesoft/agent/eventtracking/splunk/EventTrackingSplunkInternalHandlerTest.java
@@ -6,7 +6,7 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
-import com.mulesoft.agent.AgentEnableOperationException;
+import com.mulesoft.agent.exception.AgentEnableOperationException;
 import com.mulesoft.agent.domain.tracking.AgentTrackingNotification;
 
 import org.junit.Ignore;
