+++ mule-agent-gw-httpevents-internalhandlers/mule-agent-log-gw-http-tracker/src/test/java/com/mulesoft/agent/gw/http/log/GatewayHttpEventsLogInternalHandlerTest.java
@@ -3,7 +3,7 @@ package com.mulesoft.agent.gw.http.log;
 import java.util.ArrayList;
 import java.util.List;
 
-import com.mulesoft.agent.AgentEnableOperationException;
+import com.mulesoft.agent.exception.AgentEnableOperationException;
 import com.mulesoft.agent.handlers.exception.InitializationException;
 import com.mulesoft.module.client.model.HttpEvent;
