+++ mule-agent-monitoring-publishers/mule-agent-abstract-publisher/src/main/java/com/mulesoft/agent/monitoring/publisher/AbstractMonitorPublisher.java
@@ -8,7 +8,7 @@
 
 package com.mulesoft.agent.monitoring.publisher;
 
-import com.mulesoft.agent.AgentEnableOperationException;
+import com.mulesoft.agent.exception.AgentEnableOperationException;
 import com.mulesoft.agent.configuration.Configurable;
 import com.mulesoft.agent.configuration.PostConfigure;
 import com.mulesoft.agent.handlers.InternalMessageHandler;
