+++ mule-agent-monitoring-publishers/mule-agent-cloudhub-publisher/src/main/java/com/mulesoft/agent/monitoring/publisher/CloudhubMemoryMetricPublisher.java
@@ -6,7 +6,7 @@ import javax.inject.Inject;
 import javax.inject.Named;
 import javax.inject.Singleton;
 
-import com.mulesoft.agent.AgentEnableOperationException;
+import com.mulesoft.agent.exception.AgentEnableOperationException;
 import com.mulesoft.agent.buffer.BufferedHandler;
 import com.mulesoft.agent.configuration.Configurable;
 import com.mulesoft.agent.configuration.PostConfigure;
