+++ mule-agent-monitoring-publishers/mule-agent-ingest-publisher/src/main/java/com/mulesoft/agent/monitoring/publisher/ingest/IngestApplicationMonitorPublisher.java
@@ -18,7 +18,7 @@ import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
 import com.google.common.util.concurrent.ThreadFactoryBuilder;
-import com.mulesoft.agent.AgentEnableOperationException;
+import com.mulesoft.agent.exception.AgentEnableOperationException;
 import com.mulesoft.agent.configuration.Configurable;
 import com.mulesoft.agent.configuration.NotAvailableOn;
 import com.mulesoft.agent.configuration.PostConfigure;
